Change Handover

+++ handover.py
@@ -17,7 +17,7 @@ def handover_a(f_line, avg_power, hister, target):
         return 1
     else:
         dist_s0 = fabs(avg_power) + fabs(hister) - fabs(target)
-        dist_nx = fabs(target) - fabs(int(f_line[3]))
+        dist_nx = fabs(target) - fabs(float(f_line[3]))
         if fabs(dist_nx) < fabs(dist_s0):
             return 3
         else:
